test: extract child-run termination suite
